Move FileData out of Remora.Discord.Core.
