+++ Backend/Remora.Discord.Caching/API/CachingDiscordRestChannelAPI.cs
@@ -30,6 +30,7 @@ using JetBrains.Annotations;
 using Microsoft.Extensions.Options;
 using OneOf;
 using Remora.Discord.API.Abstractions.Objects;
+using Remora.Discord.API.Abstractions.Rest;
 using Remora.Discord.Caching.Services;
 using Remora.Discord.Core;
 using Remora.Discord.Rest;
